refactor: handlePlayPause 중복 호출 방지 로직을 useChannelPlayback 훅 내부로 통합
