+++ src/components/MiniPlayer.jsx
@@ -1,5 +1,3 @@
-import { useRef } from "react";
-
 import CloseIcon from "@/assets/svgs/icon-close.svg?react";
 import PauseIcon from "@/assets/svgs/icon-mini-pause.svg?react";
 import PlayIcon from "@/assets/svgs/icon-mini-play.svg?react";
@@ -28,20 +26,6 @@ const MiniPlayer = () => {
     stopWhisperServer(userId);
   };
 
-  const isProcessing = useRef(false);
-
-  const handlePlayPauseOnce = async () => {
-    if (isProcessing.current) {
-      return;
-    }
-
-    isProcessing.current = true;
-
-    await handlePlayPause();
-
-    isProcessing.current = false;
-  };
-
   return (
     <div className="absolute bottom-0 flex h-20 w-full justify-between rounded-t-4xl bg-white px-4 shadow-[0_-6px_9px_rgba(0,0,0,0.3)]">
       <div className="flex items-center">
@@ -49,7 +33,7 @@ const MiniPlayer = () => {
         <p className="ml-2 text-sm font-black">{name}</p>
       </div>
       <div className="flex items-center gap-3">
-        <Button onClick={handlePlayPauseOnce}>
+        <Button onClick={handlePlayPause}>
           {!isPlaying ? (
             <PlayIcon className="h-9 w-9 cursor-pointer" />
           ) : (
